requirements.txt
Flask
mysql-connector
pytest
mysql-connector-python

